Handle chown with no user or group

+++ pkg/security/probe/accessors.go
@@ -14557,7 +14557,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.GID"}
 		}
-		e.Chown.GID = uint32(v)
+		e.Chown.GID = int64(v)
 
 		return nil
 
@@ -14579,7 +14579,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.UID"}
 		}
-		e.Chown.UID = uint32(v)
+		e.Chown.UID = int64(v)
 
 		return nil
