+++ pkg/security/secl/model/accessors.go
@@ -14038,7 +14038,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.GID"}
 		}
-		e.Chown.GID = uint32(v)
+		e.Chown.GID = int64(v)
 
 		return nil
 
@@ -14060,7 +14060,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.UID"}
 		}
-		e.Chown.UID = uint32(v)
+		e.Chown.UID = int64(v)
 
 		return nil
